Move language model exceptions
